+++ src/components/Drawer/components/DrawerHeader/index.ts
@@ -1,0 +1,2 @@
+export { DrawerHeader } from './DrawerHeader';
+export type { DrawerHeaderProps } from './DrawerHeader.types';