+++ src/components/Drawer/helpers/DrawerOptionalHeader.tsx
@@ -1,0 +1,12 @@
+import { DrawerHeader } from '../components/DrawerHeader';
+import type { DrawerOptionalHeaderProps } from '../components/DrawerHeader/DrawerHeader.types';
+
+export const DrawerOptionalHeader = (props: DrawerOptionalHeaderProps) => {
+  const { showHeader, ...headerProps } = props;
+
+  if (!showHeader) {
+    return null;
+  }
+
+  return <DrawerHeader {...headerProps} />;
+};