+++ src/components/Icon/Icon.const.ts
@@ -1,1 +1,0 @@
-export const I_C_O_N_ROOT_CLASS = 'Bear-Icon';